Rename structs so as to not collide with class usage
