+++ Dalamud/Dalamud.cs
@@ -7,10 +7,10 @@ using System.Threading;
 using Dalamud.Configuration.Internal;
 using Dalamud.Data;
 using Dalamud.Game;
-using Dalamud.Game.Addon;
 using Dalamud.Game.ClientState;
 using Dalamud.Game.Command;
 using Dalamud.Game.Internal;
+using Dalamud.Game.Internal.Gui.Addons;
 using Dalamud.Game.Network;
 using Dalamud.Game.Text.SeStringHandling;
 using Dalamud.Hooking.Internal;
